Dedupe telemetry-explorer SQL helpers and simplify errors UI
- Extract resourceAttribute/resourceAttributeKeyExists into shared sql/
- Consolidate validateTableName behind a single module via re-exports
- Run getIssue summary and occurrences queries in parallel
- Share getErrorTraceWindow between the panel and the desktop adapter
- Drop redundant filterValue rebuild and occurrence-key guards
