+++ packages/telemetry-explorer/src/errors/index.ts
@@ -24,3 +24,4 @@ export {
   ErrorTracePanel,
   type RenderTraceLink,
 } from "./ui/error-trace-panel";
+export { getErrorTraceWindow } from "./ui/trace-window";
