+++ packages/telemetry-explorer/src/errors/sql/table.ts
@@ -1,7 +1,1 @@
-const TABLE_NAME_RE = /^[a-zA-Z_][a-zA-Z0-9_.]*$/;
-
-export function validateTableName(name: string): void {
-  if (!TABLE_NAME_RE.test(name)) {
-    throw new Error(`invalid table name: ${name}`);
-  }
-}
+export { validateTableName } from "../../sql/table";
