+++ packages/telemetry-explorer/src/errors/ui/error-trace-panel.tsx
@@ -2,14 +2,11 @@ import { Badge } from "@everr/ui/components/badge";
 import { Button } from "@everr/ui/components/button";
 import { Skeleton } from "@everr/ui/components/skeleton";
 import { formatDuration } from "@everr/ui/lib/formatting";
-import { toClickHouseDateTime } from "@everr/ui/lib/time-range";
-import { parseTimestampAsUTC } from "@everr/ui/lib/timestamp";
 import { cn } from "@everr/ui/lib/utils";
 import { ExternalLink, RefreshCw } from "lucide-react";
 import type { ReactNode } from "react";
 import type { ErrorOccurrence, RelatedSpan } from "../data/types";
-
-const TRACE_WINDOW_MS = 5 * 60 * 1000;
+import { getErrorTraceWindow } from "./trace-window";
 
 export type RenderTraceLink = (input: {
   traceId: string;
@@ -18,14 +15,6 @@ export type RenderTraceLink = (input: {
   end: string;
   children: ReactNode;
 }) => ReactNode;
-
-function getTraceWindow(timestamp: string): { start: string; end: string } {
-  const parsed = parseTimestampAsUTC(timestamp) ?? new Date();
-  return {
-    start: toClickHouseDateTime(new Date(parsed.getTime() - TRACE_WINDOW_MS)),
-    end: toClickHouseDateTime(new Date(parsed.getTime() + TRACE_WINDOW_MS)),
-  };
-}
 
 function spanStatusLabel(span: RelatedSpan, isErrorSpan: boolean): string {
   if (isErrorSpan) return "error span";
@@ -53,7 +42,7 @@ export function ErrorTracePanel({
 
   if (occurrence.traceId.trim().length === 0) return null;
 
-  const window = getTraceWindow(occurrence.timestamp);
+  const window = getErrorTraceWindow(occurrence.timestamp);
 
   return (
     <section className="min-w-0 rounded-md border bg-background">
