+++ packages/telemetry-explorer/src/sql/resource-attributes.ts
@@ -1,0 +1,7 @@
+export function resourceAttribute(key: string): string {
+  return `ResourceAttributes['${key}']`;
+}
+
+export function resourceAttributeKeyExists(key: string): string {
+  return `mapContains(ResourceAttributes, '${key}')`;
+}
